usb: dwc3: msm: Skip HSPHY suspend if already in L2 state
During dwc3_msm_prepare_suspend(), the function attempts to
suspend the HSPHY by writing to GUSB2PHYCFG to trigger an
L2 transition. However, if the HSPHY is already in L2,
this transition will not occur.

Adding a check to determine whether the HSPHY is already in L2 before
attempting the suspend operation, thereby avoiding unnecessary register
writes and ensuring correct suspend behavior.

Change-Id: I2996c00c4c45fd66958092d157b39f5b61b67b0b
Signed-off-by: Rajashekar kuruva <quic_kuruva@quicinc.com>
Signed-off-by: Prashanth K <quic_prashk@quicinc.com>
Signed-off-by: Faisal Hassan <faisal.hassan@oss.qualcomm.com>